backend/requirements.txt
fastapi
httpx
pydantic
uvicorn
sqlalchemy
python-dateutil
python-multipart
lxml
pytest
pytest-asyncio
pytest-mock
